+++ getl/common/upsert.py
@@ -1,10 +1,11 @@
 from contextlib import ExitStack, contextmanager
 from dataclasses import InitVar, dataclass, field
 from functools import partial
-from itertools import islice
-from typing import Any, Callable, Iterable, Iterator, List, Optional
+from operator import itemgetter
+from typing import Any, Callable, Iterable, List, Optional
 
 import psycopg2
+from psycopg2.extras import execute_values
 from psycopg2.sql import SQL, Composed, Identifier, Placeholder
 
 
@@ -49,8 +50,9 @@ def handle_partition(
             update_columns=update_columns,
         )
 
-        for rows in chunked(iterator, 1000):
-            puq.execute(cursor, rows, columns_map=columns)
+        rows = map(itemgetter(*columns), iterator)
+
+        puq.execute(cursor, rows, page_size=1000)
 
 
 @contextmanager
@@ -59,15 +61,6 @@ def postgres_connection_cursor(*args, **kwargs):
         conn = stack.enter_context(psycopg2.connect(*args, **kwargs))
         cursor = stack.enter_context(conn.cursor())
         yield cursor
-
-
-def chunked(iterator: Iterable[Any], size) -> Iterator[List[Any]]:
-    iterator_ = iter(iterator)
-    while True:
-        rows = list(islice(iterator_, size))
-        if not rows:
-            break
-        yield rows
 
 
 @dataclass
@@ -81,7 +74,6 @@ class PostgresUpsertQuery:
     sql_columns: Composed = field(init=False)
     sql_conflict_columns: Composed = field(init=False)
     sql_update: Composed = field(init=False)
-    sql_base_fields: Composed = field(init=False, repr=False)
 
     def __post_init__(self, table, columns, conflict_columns, update_columns):
         # Inits the base sql statements to be build later once the rows are available
@@ -98,33 +90,21 @@ class PostgresUpsertQuery:
             for col in update_columns
         )
 
-        # sql_base_fields creates sql placeholders for the columns
-        # Example, 2 columns will generate:
-        # (%s, %s)
-        self.sql_base_fields = SQL("({})").format(
-            SQL(", ").join(Placeholder() * len(columns))
-        )
-
-    def build_query(self, len_rows: int) -> SQL:
+    def build_query(self) -> SQL:
         sql_query = """
         INSERT INTO {table} ({columns})
-        VALUES {fields}
+        VALUES {values}
         ON CONFLICT ({conflict_columns})
         DO UPDATE SET {update}
         """
         return SQL(sql_query).format(
             table=self.sql_table,
             columns=self.sql_columns,
+            values=Placeholder(),
             conflict_columns=self.sql_conflict_columns,
             update=self.sql_update,
-            fields=SQL(", ").join(self.sql_base_fields * len_rows),
         )
 
-    def execute(self, cursor, rows: List[Any], columns_map: List[str]) -> None:
-        sql_query = self.build_query(len(rows))
-        flattened_rows = flatten_rows_dict(rows, columns_map)
-        cursor.execute(sql_query, flattened_rows)
-
-
-def flatten_rows_dict(rows: List[dict], columns: List[str]):
-    return [row[col] for row in rows for col in columns]
+    def execute(self, cursor, rows: Iterable[Any], page_size: int = 100) -> None:
+        sql_query = self.build_query()
+        execute_values(cursor, sql_query, rows, page_size=page_size)